Refactored variable names
